# openapi.yaml
openapi: 3.1.0
info:
  title: SEMD API
  version: 1.0.0
paths:
...
  /ml/service:
    get:
      tags:
      - ml
      summary: Get Service
      operationId: get_service_ml_service_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MLServiceResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
...
  /setting/third-service/:
    get:
      tags:
      - third-service
      summary: List Third-Party Services
      description: List all third-party service configurations for the current user
      operationId: list_third_services_setting_third_service__get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ThirdServiceResponse'
                title: Response List Third Services Setting Third Service  Get
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    post:
      tags:
      - third-service
      summary: Create Third-Party Service
      description: Create a new third-party service configuration
      operationId: create_third_service_setting_third_service__post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThirdServiceCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThirdServiceResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /setting/third-service/{id}:
    get:
      tags:
      - third-service
      summary: Get Third-Party Service
      description: Retrieve a third-party service configuration by ID
      operationId: get_third_service_setting_third_service__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThirdServiceResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    put:
      tags:
      - third-service
      summary: Update Third-Party Service
      description: Update a third-party service configuration
      operationId: update_third_service_setting_third_service__id__put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThirdServiceUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThirdServiceResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    delete:
      tags:
      - third-service
      summary: Delete Third-Party Service
      description: Soft delete a third-party service (sets is_active=False)
      operationId: delete_third_service_setting_third_service__id__delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /setting/third-service/{id}/execute:
    post:
      tags:
      - third-service
      summary: Execute Third-Party Service
      description: Execute a third-party service with runtime variables
      operationId: execute_third_service_setting_third_service__id__execute_post
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThirdServiceExecuteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Execute Third Service Setting Third Service  Id  Execute
                  Post
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /setting/third-service/{id}/rest-api/test:
    post:
      tags:
      - third-service
      summary: Test Third Service
      description: Test a third-party service API
      operationId: test_third_service_setting_third_service__id__rest_api_test_post
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThirdServiceExecuteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Test Third Service Setting Third Service  Id  Rest
                  Api Test Post
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /setting/service/:
    get:
      tags:
      - service-configuration
      summary: List Service Configurations
      description: Get all service configurations for the current user
      operationId: list_services_setting_service__get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceConfModel'
                title: Response List Services Setting Service  Get
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /setting/service/{service_id}:
    get:
      tags:
      - service-configuration
      summary: Get Service Configuration
      description: Get a specific service configuration by ID
      operationId: get_service_setting_service__service_id__get
      parameters:
      - name: service_id
        in: path
        required: true
        schema:
          type: integer
          title: Service Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceConfModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
...
  /setting/system-config:
    get:
      tags:
      - system-config
      summary: '[Admin] Get All System Configs'
      description: Get all system configuration values. Admin only.
      operationId: get_all_configs_setting_system_config_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigListResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /setting/system-config/{config_key}:
    get:
      tags:
      - system-config
      summary: '[Admin] Get System Config by Key'
      description: Get a specific system configuration by key. Admin only.
      operationId: get_config_by_key_setting_system_config__config_key__get
      parameters:
      - name: config_key
        in: path
        required: true
        schema:
          type: string
          title: Config Key
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - system-config
      summary: '[Admin] Update System Config'
      description: Update a system configuration value by key. Admin only.
      operationId: update_config_setting_system_config__config_key__put
      parameters:
      - name: config_key
        in: path
        required: true
        schema:
          type: string
          title: Config Key
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfigUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /stat/third-party/services:
    get:
      tags:
      - stat-third-party
      summary: Get Third Party Services
      operationId: get_third_party_services_stat_third_party_services_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThirdPartyServiceResponse'
        '404':
          description: Not found
        '501':
          description: Not implemented
        '422':
          description: Validation error
...
components:
  schemas:
...
    BaseResponseModel:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
      type: object
      required:
      - status
      - message
      title: BaseResponseModel
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MLServiceResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          title: data
          description: ML service data
      type: object
      required:
      - status
      - message
      - data
      title: MLServiceResponse
...
    ServiceConfModel:
      properties:
        service_conf_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Service Conf Id
        user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: User Id
        service_name:
          type: string
          maxLength: 32
          title: Service Name
        service_type:
          $ref: '#/components/schemas/ServiceType'
        is_active:
          type: boolean
          title: Is Active
          default: true
        version_no:
          type: string
          maxLength: 12
          title: Version No
        config_uri:
          type: string
          title: Config Uri
        config_json:
          additionalProperties: true
          type: object
          title: Config Json
          default: {}
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - service_name
      - service_type
      - version_no
      - config_uri
      title: ServiceConfModel
    ServiceType:
      type: string
      enum:
      - ML_MODEL
      - REST_API
      title: ServiceType
    SystemConfigListResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/SystemConfigModel'
          type: array
          title: data
          description: List of system configs
      type: object
      required:
      - status
      - message
      - data
      title: SystemConfigListResponse
    SystemConfigModel:
      properties:
        system_config_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: System Config Id
        config_key:
          type: string
          title: Config Key
        config_value:
          type: string
          title: Config Value
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - config_key
      - config_value
      title: SystemConfigModel
    SystemConfigResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          $ref: '#/components/schemas/SystemConfigModel'
          title: data
          description: System config data
      type: object
      required:
      - status
      - message
      - data
      title: SystemConfigResponse
    SystemConfigUpdateRequest:
      properties:
        config_value:
          type: string
          title: Config Value
          description: New value for the config
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Optional description update
      type: object
      required:
      - config_value
      title: SystemConfigUpdateRequest
...
    ThirdPartyServiceItem:
      properties:
        service_id:
          type: string
          title: Service Id
          description: Service ID
        service_name:
          type: string
          title: Service Name
          description: Service name
        service_type:
          type: string
          title: Service Type
          description: 'Service type: url_scanner, threat_intel, etc'
        status:
          type: string
          title: Status
          description: 'Service status: active, inactive, error'
        request_count:
          type: integer
          title: Request Count
          description: Number of requests
        success_rate:
          type: number
          title: Success Rate
          description: Success rate percentage
        avg_response_time:
          type: number
          title: Avg Response Time
          description: Average response time in ms
        last_used:
          type: string
          title: Last Used
          description: Last used timestamp
      type: object
      required:
      - service_id
      - service_name
      - service_type
      - status
      - request_count
      - success_rate
      - avg_response_time
      - last_used
      title: ThirdPartyServiceItem
    ThirdPartyServiceResponse:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
        data:
          items:
            $ref: '#/components/schemas/ThirdPartyServiceItem'
          type: array
          title: Data
      type: object
      required:
      - status
      - message
      - data
      title: ThirdPartyServiceResponse
...
    ThirdServiceCreateRequest:
      properties:
        service_name:
          type: string
          maxLength: 64
          title: Service Name
          example: Cloudflare URL Scanner
        base_url:
          type: string
          title: Base Url
          description: Base URL with optional {{variable}} placeholders
          example: https://api.cloudflare.com/client/v4/accounts/{{account_id}}/urlscanner/v2/scan
        http_method:
          type: string
          maxLength: 8
          title: Http Method
          default: POST
        headers_json:
          additionalProperties:
            type: string
          type: object
          title: Headers Json
          description: HTTP headers including authentication (e.g., Authorization,
            API keys)
          example:
            Authorization: Bearer your_token_here
            Content-Type: application/json
        config_json:
          additionalProperties: true
          type: object
          title: Config Json
          description: Configuration for body template and URL template
          example:
            body_template:
            - input: url
              key: url
            response_mapping:
              is_malicious: result.malicious
            url_template:
              path_params:
                account_id: your_account_id
              query_params:
                url: url
        mapping_json:
          additionalProperties: true
          type: object
          title: Mapping Json
          description: Response mapping configuration for extracting prediction data
          example:
            class: data.category
            is_malicious: data.is_malicious
            suggested_desc: data.description
      type: object
      required:
      - service_name
      - base_url
      title: ThirdServiceCreateRequest
    ThirdServiceExecuteRequest:
      properties:
        vars:
          additionalProperties: true
          type: object
          title: Vars
          description: Runtime variables to resolve body template
      type: object
      required:
      - vars
      title: ThirdServiceExecuteRequest
    ThirdServiceResponse:
      properties:
        third_service_conf_id:
          type: integer
          title: Third Service Conf Id
        service_conf_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Service Conf Id
        service_name:
          type: string
          title: Service Name
        base_url:
          type: string
          title: Base Url
        http_method:
          type: string
          title: Http Method
        headers_json:
          additionalProperties: true
          type: object
          title: Headers Json
        config_json:
          additionalProperties: true
          type: object
          title: Config Json
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - third_service_conf_id
      - service_name
      - base_url
      - http_method
      - headers_json
      - config_json
      - is_active
      - created_at
      - updated_at
      title: ThirdServiceResponse
    ThirdServiceUpdateRequest:
      properties:
        service_name:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Service Name
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        http_method:
          anyOf:
          - type: string
            maxLength: 8
          - type: 'null'
          title: Http Method
        headers_json:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Headers Json
        config_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Config Json
        mapping_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Mapping Json
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      type: object
      title: ThirdServiceUpdateRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError